requirements.txt
fastapi
uvicorn
python-multipart
pydantic
passlib
bcrypt
python-dotenv
streamlit
httpx